Infer matmul dimension roles to compute vectorization (#2303)
This PR does the following:
1. Rename `RolesMap` to `TensorRolesMap` and introduce `DimRolesMap`
which is a mapping from `ValGroup` to `MatmulDomain`.
2. Compute a canonical dim ordering on `ValGroup`s based on allocation
domains of inputs and outputs. This is used to compute vectorization
properly but can be used for canonicalization of loop domains in
scheduleMatmul in a future PR.
3. Properly infer vectorization for every operand, epilogue input, and
output based on canonical dim ordering.

This is in preparation for further generalization to accomodate multiple
MmaOps in a single Fusion.

Fixes #2169.

---------

+++ tests/cpp/test_gpu_tensorcore.cpp
@@ -3092,7 +3092,7 @@ TEST_F(GPUTTensorCoreTest, MisalignedVectorization) {
                  {504, 136, 248, 8, 8, 2}, // epilogue not vectorizable due to
                  // offset
              }) {
-          const auto maybeUnalign = [](const at::Tensor& t, int offset) {
+          const auto maybeUnalign = [](const at::Tensor& t, int64_t offset) {
             if (offset == 16 / t.element_size()) {
               // Already fully aligned
               return t;
